pyproject.toml
[project]
name = "msw-translation"
version = "0.1.0"
description = "Real-time game overlay translator — OCR → translate → transparent overlay"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6",
    "Pillow>=10.0",
    # Cross-platform screenshot (Windows / Linux)
    "mss>=9.0; sys_platform != 'darwin'",
    # OCR fallback (Linux only — Windows uses Windows.Media.Ocr, macOS uses Apple Vision)
    "pytesseract>=0.3; sys_platform == 'linux'",
    # Windows WinRT bindings for OCR + translation
    "winrt-runtime>=2.0; sys_platform == 'win32'",
    "winrt-Windows.Media.Ocr>=2.0; sys_platform == 'win32'",
    "winrt-Windows.Globalization>=2.0; sys_platform == 'win32'",
    "winrt-Windows.Graphics.Imaging>=2.0; sys_platform == 'win32'",
]

[project.optional-dependencies]
google = ["googletrans==4.0.0rc1"]
dev = ["pytest>=8.0", "pyinstaller>=6.0"]

[project.scripts]
msw-translation = "main:main"

[tool.uv]
dev-dependencies = ["pytest>=8.0", "pyinstaller>=6.0"]
